+++ src/video_rental/Price.java
@@ -2,5 +2,5 @@ package video_rental;
 
 public abstract class Price {
 	abstract double getCharge(int daysRented);
-	abstract double getFrequentPoints(int daysRented);
+	abstract int getFrequentPoints(int daysRented);
 }
